requirements.txt
kafka-python
streamlit
